backend/requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
langchain
langchain-qdrant
langchain-openai
openai
qdrant-client
pydantic
pypdf
tiktoken
